requirements.txt
fastapi
pyrebase4